Refactor convert_str_and_pathlib_paths to stop weird call signature.
