+++ datashuttle/configs/config_class.py
@@ -16,7 +16,11 @@ from pathlib import Path
 
 import yaml
 
-from datashuttle.configs import canonical_configs, canonical_folders
+from datashuttle.configs import (
+    canonical_configs,
+    canonical_folders,
+    load_configs,
+)
 from datashuttle.utils import folders, utils
 
 
@@ -52,18 +56,13 @@ class Configs(UserDict):
         self.project_name = project_name
         self.file_path = file_path
 
-        self.keys_str_on_file_but_path_in_class = [
-            "local_path",
-            "central_path",
-        ]
-
         self.logging_path: Path
         self.hostkeys_path: Path
         self.ssh_key_path: Path
         self.project_metadata_path: Path
 
     def setup_after_load(self) -> None:
-        self.convert_str_and_pathlib_paths(self, "str_to_path")
+        load_configs.convert_str_and_pathlib_paths(self, "str_to_path")
         self.ensure_local_and_central_path_end_in_project_name()
         self.check_dict_values_raise_on_fail()
 
@@ -109,7 +108,7 @@ class Configs(UserDict):
         Save the dictionary to .yaml file stored in self.file_path.
         """
         cfg_to_save = copy.deepcopy(self.data)
-        self.convert_str_and_pathlib_paths(cfg_to_save, "path_to_str")
+        load_configs.convert_str_and_pathlib_paths(cfg_to_save, "path_to_str")
 
         with open(self.file_path, "w") as config_file:
             yaml.dump(cfg_to_save, config_file, sort_keys=False)
@@ -123,47 +122,13 @@ class Configs(UserDict):
         with open(self.file_path, "r") as config_file:
             config_dict = yaml.full_load(config_file)
 
-        self.convert_str_and_pathlib_paths(config_dict, "str_to_path")
+        load_configs.convert_str_and_pathlib_paths(config_dict, "str_to_path")
 
         self.data = config_dict
 
     # -------------------------------------------------------------------------
     # Utils
     # -------------------------------------------------------------------------
-
-    # TODO: all of this should move to canonical configs!!!
-    def convert_str_and_pathlib_paths(
-        self, config_dict: Union["Configs", dict], direction: str
-    ) -> None:
-        """
-        Config paths are stored as str in the .yaml but used as Path
-        in the module, so make the conversion here.
-
-        Parameters
-        ----------
-
-        config_dict : DataShuttle.cfg dict of configs
-        direction : "path_to_str" or "str_to_path"
-        """
-        for path_key in self.keys_str_on_file_but_path_in_class:
-            value = config_dict[path_key]
-
-            if path_key == "central_path" and value is None:
-                continue
-
-            if value:
-                if direction == "str_to_path":
-                    config_dict[path_key] = Path(value)
-
-                elif direction == "path_to_str":
-                    if not isinstance(value, str):
-                        config_dict[path_key] = value.as_posix()
-
-                else:
-                    utils.log_and_raise_error(
-                        "Option must be 'path_to_str' or 'str_to_path'",
-                        ValueError,
-                    )
 
     def build_project_path(
         self,
